requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
defusedxml==0.4.1
dj-database-url==0.4.0
Django==1.9.1
django-allauth==0.24.1
django-widget-tweaks==1.3
gunicorn==19.4.5
kombu==3.0.26
oauthlib==1.0.3
pip-autoremove==0.9.0
psycopg2==2.6.1
python3-openid==3.0.9
pytz==2015.4
requests==2.9.1
requests-oauthlib==0.6.0
vk==2.0a4
wheel==0.24.0
whitenoise==2.0.6
